+++ client/app-src/ui/converters/DataInvalidaException.js
@@ -1,0 +1,8 @@
+import { ApplicationException } from '../../util/ApplicationException';
+
+export class DataInvalidaException extends ApplicationException {
+    constructor() {
+
+        super('A data dave estar no formato dd/mm/aaaa');
+    }
+}
