+++ client/app-src/ui/views/MensagemView.js
@@ -1,0 +1,10 @@
+import { View } from './View';
+
+export class MensagemView extends View {
+
+    template(model) {
+        return model.texto
+            ? `<p class="alert alert-info">${model.texto}</p>`
+            : `<p></p>`;
+    }
+}
