+++ client/app-src/ui/views/NegociacoesView.js
@@ -1,0 +1,40 @@
+import { View } from './View';
+import { DateConverter } from '../converters/DateConverter'; 
+
+export class NegociacoesView extends View {
+
+    template(model) {
+        return `
+        <table class="table table-hover table-bordered">
+            <thead>
+                <tr>
+                    <th>DATA</th>
+                    <th>QUANTIDADE</th>
+                    <th>VALOR</th>
+                    <th>VOLUME</th>
+                </tr>
+            </thead>
+            
+            <tbody>
+                ${model.paraArray().map(negociacao => 
+                    `
+                    <tr>
+                        <td>${ DateConverter.paraTexto(negociacao.data) }</td>
+                        <td>${ negociacao.quantidade }</td>
+                        <td>${ negociacao.valor }</td>
+                        <td>${ negociacao.volume }</td>
+                    </tr>
+                    `
+                ).join('')}
+            </tbody>
+            
+            <tfoot>
+                <tr>
+                    <td colspan="3"></td>
+                    <td>${model.volumeTotal}</td>
+                </tr>
+            </tfoot>
+        </table>
+        `;
+    }
+}
