+++ client/app-src/util/ApplicationException.js
@@ -1,0 +1,7 @@
+export class ApplicationException extends Error {
+    constructor(msg = '') {
+
+        super(msg);
+        this.name = this.constructor.name;
+    }
+}
