+++ client/app-src/util/Bind.js
@@ -1,0 +1,13 @@
+import { ProxyFactory } from './ProxyFactory';
+
+export class Bind {
+    constructor(model, view, ...props) {
+        const proxy = ProxyFactory.create(model, props, model => {
+            view.update(model);
+        });
+
+        view.update(model);
+
+        return proxy;
+    }
+}
