+++ client/app-src/util/DaoFactory.js
@@ -1,0 +1,9 @@
+import { ConnectionFactory } from './ConnectionFactory';
+import { NegociacaoDao } from '../domain/negociacao/NegociacaoDao';
+
+export function getNegociacaoDao() {
+    
+    return ConnectionFactory
+            .getConnection()
+            .then(conn => new NegociacaoDao(conn));
+}
