--- client/app/domain/negociacao/Negociacao.js
@@ -1,30 +1,0 @@
-class Negociacao {
-    constructor(_data, _quantidade, _valor) {
-        Object.assign(this, {_quantidade, _valor });
-        this._data = new Date(_data.getTime());
-        Object.freeze(this);
-        // this._quantidade = quantidade;
-        // this._valor = valor;
-    }
-    
-    get volume() {
-        return this._quantidade * this._valor;
-    }
-
-    get data() {
-        return new Date(this._data.getTime());
-    }
-
-    get quantidade() {
-        return this._quantidade;
-    }
-
-    get valor() {
-        return this._valor;
-    }
-
-    equals(negociacao) {
-
-        return JSON.stringify(this) == JSON.stringify(negociacao);        
-    }
-}
