--- client/app/ui/converters/DataInvalidaException.js
@@ -1,6 +1,0 @@
-class DataInvalidaException extends ApplicationException {
-    constructor() {
-
-        super('A data dave estar no formato dd/mm/aaaa');
-    }
-}
