--- client/app/ui/models/Mensagem.js
@@ -1,13 +1,0 @@
-class Mensagem {
-    constructor(texto = '') {
-        this._texto = texto;
-    }
-
-    get texto() {
-        return this._texto;
-    }
-
-    set texto(texto) {
-        this._texto = texto;
-    }
-}
