--- client/app/ui/views/MensagemView.js
@@ -1,8 +1,0 @@
-class MensagemView extends View {
-
-    template(model) {
-        return model.texto
-            ? `<p class="alert alert-info">${model.texto}</p>`
-            : `<p></p>`;
-    }
-}
