--- client/app/ui/views/View.js
@@ -1,13 +1,0 @@
-class View {
-    constructor(seletor) {
-        this._elemento = document.querySelector(seletor);
-    }
-
-    update(model) {
-        this._elemento.innerHTML = this.template(model);
-    }
-
-    template(model) {
-        throw new Error('Você precisa implementar o método template');
-    }
-}
