--- client/app/util/ApplicationException.js
@@ -1,7 +1,0 @@
-class ApplicationException extends Error {
-    constructor(msg = '') {
-
-        super(msg);
-        this.name = this.constructor.name;
-    }
-}
