--- client/app/util/Bind.js
@@ -1,11 +1,0 @@
-class Bind {
-    constructor(model, view, ...props) {
-        const proxy = ProxyFactory.create(model, props, model => {
-            view.update(model);
-        });
-
-        view.update(model);
-
-        return proxy;
-    }
-}
