--- client/app/util/DaoFactory.js
@@ -1,6 +1,0 @@
-function getNegociacaoDao() {
-    
-    return ConnectionFactory
-            .getConnection()
-            .then(conn => new NegociacaoDao(conn));
-}
