+++ client/index.html
@@ -46,24 +46,12 @@
     
     <div id="negociacoes"></div>
 
-    <script src="app/domain/negociacao/Negociacao.js"></script>
-    <script src="app/controllers/NegociacaoController.js"></script>
-    <script src="app/ui/converters/DateConverter.js"></script>
-    <script src="app/domain/negociacao/Negociacoes.js"></script>
-    <script src="app/ui/views/View.js"></script>
-    <script src="app/ui/views/NegociacoesView.js"></script>
-    <script src="app/ui/models/Mensagem.js"></script>
-    <script src="app/ui/views/MensagemView.js"></script>
-    <script src="app/util/ProxyFactory.js"></script>
-    <script src="app/util/Bind.js"></script>
-    <script src="app/util/ApplicationException.js"></script>
-    <script src="app/ui/converters/DataInvalidaException.js"></script>
-    <script src="app/domain/negociacao/NegociacaoService.js"></script>
-    <script src="app/util/HttpService.js"></script>
-    <script src="app/util/ConnectionFactory.js"></script>
-    <script src="app/domain/negociacao/NegociacaoDao.js"></script>
-    <script src="app/util/DaoFactory.js"></script>
-    <script src="app/app.js"></script>
+    <script src="node_modules/systemjs/dist/system.js"></script>
+    <script>
+        System
+            .import('app/app.js')
+            .catch(err => console.error(err));
+    </script>
 
 </body>
 </html>
